% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STests.R
\name{ARIMAWrapper2}
\alias{ARIMAWrapper2}
\title{Estimates autoregression coefficient for a vector, wrapper}
\usage{
ARIMAWrapper2(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Autoregression coefficient value mainly a number between zero and unity
}
\description{
Wrapper for autoregression coefficient calculator. Works on any type of input data.
}
\author{
Harutyun Khachatryan
}
