% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STests.R
\name{ATest2}
\alias{ATest2}
\title{Performs stationarity change test based on autoregression coefficient}
\usage{
ATest2(itseries, ind, lobs)
}
\arguments{
\item{itseries}{time series like vector}

\item{ind}{index around which test is performed}

\item{lobs}{length of window around ind index}
}
\value{
0 when nor mean neither variance change is detected,
0.1 when both mean and variance change is detected
}
\description{
Does stationarity change test around index ind of time series with window length lobs
}
\author{
Harutyun Khachatryan, Tigran Khachikyan
}
