% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stationarity.R
\name{BadInd2}
\alias{BadInd2}
\title{Bad data finder}
\usage{
BadInd2(intseries, norm_dates, real_dates, stat_changes, xhint = FALSE)
}
\arguments{
\item{intseries}{time series like vector}

\item{norm_dates}{timeseries observation times given as vector of POSIXct}

\item{real_dates}{real observation times given as vector of POSIXct}

\item{stat_changes}{indeces of stationarity changes in time series}

\item{xhint}{logical TRUE/FALSE indicating if data is highly unstable (default value is FALSE)}
}
\description{
Finds locations of bad days in data according to stationarity change points
}
\author{
Harutyun Khachatryan
}
