% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divider.R
\name{CalcOnGroups2}
\alias{CalcOnGroups2}
\title{Does calculation on groups}
\usage{
CalcOnGroups2(div_groups, func_calc, func_def = function(w) median(w, na.rm =
  TRUE))
}
\arguments{
\item{div_groups}{list of vectors on which statistical calculation should be done}

\item{func_calc}{name of statistical or other function that can be applied on vector of values}

\item{func_def}{a function name that should be used in case when func_calc gives NA values, default value is median}
}
\value{
results of calculation for each group
}
\description{
Calculates some statistics defined by functions on list of vectors and returns it as a list of vectors
}
\author{
Harutyun Khachatryan
}
