% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divider.R
\name{DivGroups2}
\alias{DivGroups2}
\title{Projects all data to one day and divides data into groups}
\usage{
DivGroups2(time_stamps, values, delta_min, ch_ind = NA, overlap = TRUE,
  loop = TRUE)
}
\arguments{
\item{time_stamps}{one dimensional array of timestamp}

\item{values}{one dimensional array of values measured at times 'time_stamps', i.e. they should have the same length}

\item{delta_min}{time aggregation parameter given in minutes}

\item{ch_ind}{integer value of index at which concept changes last occur}

\item{overlap}{logical TRUE/FALSE, (default TRUE) indicating if derived groups will be overlapping or not}

\item{loop}{logical TRUE/FALSE, (default TRUE) indicating if last and first groups will be mirrored or not}
}
\value{
list of vectors representing values in groups, name of each list is timestamps around which values appeared in one day projection
}
\description{
Takes timestamps and values and creates groups according to projected timestamp position in day
}
\note{
Monitis
}
\author{
Harutyun Khachatryan
}
\keyword{series}
\keyword{time}
