% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodicity.R
\name{DRel2}
\alias{DRel2}
\title{Calculates main periodicity markers for itseries}
\usage{
DRel2(itseries, nfreq)
}
\arguments{
\item{itseries}{timeseries like vector of values}

\item{nfreq}{number of steps of possible periodicity to be checked}
}
\value{
vector of 4 statistical parameters showing periodicity,
a) last period seasonal component and remainder component standard deviation ratio,
b) Spearman correlation rho for last period and one before last one
c) Spearman correlation rho for last period and mean periodic pattern of data
d) median value of seasonal and remainder components ratio for last period
}
\description{
Estimates main statistical parameters that shows that data is periodic
}
\author{
Harutyun Khachatryan, Tigran Khachikyan
}
