% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistAvg2.R
\name{HistAvg2.cv}
\alias{HistAvg2.cv}
\title{Dynamic threshold calculator, cross-validation}
\usage{
HistAvg2.cv(json.str, figure.out = FALSE, legacy.plot = FALSE, dmin = 10,
  out.path = " ", sens = "low")
}
\arguments{
\item{json.str}{JSON input can be both string or file path}

\item{figure.out}{logical TRUE/FALSE indicating if after analysis a plot of analysis will be created}

\item{legacy.plot}{logical TRUE/FALSE indicating if plot will be new style i.e. last three days with thresholds
or old one i.e. 2 plots: 1) raw data plot and 2) all raw data points, last day line and thresholds
(default value is FALSE)}

\item{dmin}{fixed step (minutes) between two neighbours in derived dynamic threshold. Default value is 10, but
in case when it is impossible to calculate with that step, then it autocorrected}

\item{out.path}{plot output file folder path, should be accessible for write. Default value is " ", in that
case plot file will be generated in working directory of R and plot file path anyway is written in output information
generated by this method}

\item{sens}{shows confidence interval has three values "low", "medium" and "high" (default is "low")}
}
\description{
Splits data into train and test sets.
Calculates dynamic threshold on train set.
Then checks how many points violate dynamic threshold
and estimates accuracy value of dynamic threshold whether
that violation meaningfull or not.
This method calls HistAvg.test method, more expalntations
find there.
}
\author{
Harutyun Khachatryan
}
