% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistAvg2.R
\name{HistAvg2.test}
\alias{HistAvg2.test}
\title{Dynamic threshold calculator, test purposes}
\usage{
HistAvg2.test(json.str, figure.out = FALSE, legacy.plot = FALSE,
  dmin = 10, out.path = " ", sens = "low", test.mode = FALSE)
}
\arguments{
\item{json.str}{JSON input can be both string or file path}

\item{figure.out}{logical TRUE/FALSE indicating if after analysis a plot of analysis will be created}

\item{legacy.plot}{logical TRUE/FALSE indicating if plot will be new style i.e. last three days with thresholds
or old one i.e. 2 plots: 1) raw data plot and 2) all raw data points, last day line and thresholds
(default value is FALSE)}

\item{dmin}{fixed step (minutes) between two neighbours in derived dynamic threshold. Default value is 10, but
in case when it is impossible to calculate with that step, then it autocorrected}

\item{out.path}{plot output file folder path, should be accessible for write. Default value is " ", in that
case plot file will be generated in working directory of R and plot file path anyway is written in output information
generated by this method}

\item{sens}{shows confidence interval has three values "low", "medium" and "high" (default is "low")}

\item{test.mode}{logical TRUE/FALSE showing if test mode is enabled or not, default is FALSE}
}
\description{
Finds dynamic threshold for data in JSON format,
HistAvg2 wrapped in try to work without any error on MapReduce or Spark framework
This method works in any case and if something goes wrong writes output JSON
format string explaining problem
}
\author{
Harutyun Khachatryan, Tigran Khachikyan, Anoush Ghambaryan
}
