% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divider.R
\name{HistOutliers2}
\alias{HistOutliers2}
\title{Gives an array of outliers}
\usage{
HistOutliers2(div_groups, limits)
}
\arguments{
\item{div_groups}{one day projected groups of data in list of vectors}

\item{limits}{a list of vectors containing lower and upper limits}
}
\value{
list of outlisers, values shows their distance from limits
}
\description{
Finds outliers from given limits for historical data and represents every outlier by it's distance from limits
}
\author{
Harutyun Khachatryan
}
