% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divider.R
\name{HistQuantile2}
\alias{HistQuantile2}
\title{Calculates quantiles for historical data}
\usage{
HistQuantile2(x, hprobs, limits, nout)
}
\arguments{
\item{x}{historical data values}

\item{hprobs}{probs of quantiles needed for calculation, for example c(0.01,0.99)}

\item{limits}{limits on values within which most of historical values should lay but for some caes it can goes out}

\item{nout}{integer value shows how many historical values should go out limits that it can be treated as tendency}
}
\value{
lower and upper bounds for historical data
}
\description{
Gives quantiles of historical data for predefined quantiles hprobs
}
\author{
Harutyun Khachatryan
}
