% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSTest.R
\name{HRound}
\alias{HRound}
\title{My own redifintion of round function}
\usage{
HRound(x)
}
\arguments{
\item{x}{numeric value to be rounded}
}
\value{
rounded value xr of original value x
}
\description{
Redifinition of round function in cases when even number +0.5 rounded downstair
}
\author{
Harutyun Khachatryan
}
