% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STests.R
\name{Hurst2}
\alias{Hurst2}
\title{Estimates Hurst exponent for a vector}
\usage{
Hurst2(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Hurst exponent value mainly a number between zero and unity
}
\description{
Calculates Hurst exponent, i.e. long-time memory for a vector
}
\author{
Harutyun Khachatryan
}
