% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STests.R
\name{HurstWrapper2}
\alias{HurstWrapper2}
\title{Estimates Hurst exponent for a vector, wrapper}
\usage{
HurstWrapper2(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Hurst exponent value mainly a number between zero and unity
}
\description{
Wrapper for Hurst exponent calculator. Works on any type of input data.
}
\author{
Harutyun Khachatryan
}
