% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSTest.R
\name{KSInfo2}
\alias{KSInfo2}
\title{Extra information from Kolmogorov-Smirnov lambda}
\usage{
KSInfo2(lambdas)
}
\arguments{
\item{lambdas}{2-dimensional matrix of KS lambdas and indeces}
}
\value{
a list of different parameters:
1) a field named "weekchange" conatins indeces of stationarity change points affected by week pattern existance in data
2) a field named "type" has two possible values "low" and "high" indicating type of week pattern affected changes
3) a field named "nchange" number of possible stationarity changes in data
4) a field named "weekly7" logical indicating if just 7 day pattern exists
5) a field named "weekly52" logical indicating if just 5-2 (working/non-working) days pattern exists
6) a field named "highly_unstable" logical showing if data is higly unstable, can have increasing or decreasing trend
}
\description{
Finds different information from KS lambdas like: number of change points, weekly pattern, stationarity
}
\author{
Harutyun Khachatryan
}
