% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSTest.R
\name{KSValues2}
\alias{KSValues2}
\title{Estimates Kolmogorov-Smirnov's test statistic lambda}
\usage{
KSValues2(itseries, lobs)
}
\arguments{
\item{itseries}{time series like vector of values}

\item{lobs}{length of window around ind index}
}
\value{
two-dimensional matrix of indeces and lambdas
}
\description{
Gives a two-dimensional matrix of indeces and lambdas
}
\author{
Harutyun Khachatryan
}
