% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSTest.R
\name{KSWrapper2}
\alias{KSWrapper2}
\title{Estimates Kolmogorov-Smirnov's test statistic lambda, wrapper}
\usage{
KSWrapper2(itseries, ind, lobs)
}
\arguments{
\item{itseries}{time series like vector of values}

\item{ind}{index around which test is performed}

\item{lobs}{length of window around ind index}
}
\value{
Kolmogorov-Smirnov's test lambda
}
\description{
Extracts from stats ks.test method main statistic lambda, works on any input data
}
\author{
Harutyun Khachatryan
}
