% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSTest.R
\name{KTest2}
\alias{KTest2}
\title{Performs Kolmogorov-Smirnov's test}
\usage{
KTest2(itseries, ind, lobs, robust = 1)
}
\arguments{
\item{itseries}{time series like vector of values}

\item{ind}{index around which test is performed}

\item{lobs}{length of window around ind index}

\item{robust}{integer value with level 0, 1, 2 (default 1)
a) when robust is 0 then estimated KS lambda should be greater or equal than exp(1) for detecting ind as change point
b) when robust is 1 then estimated KS lambda should be greater or equal than 4.2 for detecting ind as change point
c) when robust is 2 then estimated KS lambda should be greater or equal than 7.0 for detecting ind as change point}
}
\value{
0 if no stationarity change is detected 0.1 otherwise
}
\description{
Tests if ind index is a point of stationarity change by mean of Kolmogorov-Smirnov test
}
\author{
Harutyun Khachatryan
}
