% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STests.R
\name{LDTest2}
\alias{LDTest2}
\title{Searches lost data around stationarity change points}
\usage{
LDTest2(xtimes, obstimes, lobs)
}
\arguments{
\item{xtimes}{observation times of timeseries}

\item{obstimes}{observation times of raw data}

\item{lobs}{length of window around ind index}
}
\value{
0 when change point is not affected by lost data 0.1 otherwise
}
\description{
Checks whether stationarity change appears due to lost data
}
\author{
Harutyun Khachatryan
}
