% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicter.R
\name{LinearModel2}
\alias{LinearModel2}
\title{Creates linear prediction model for given data}
\usage{
LinearModel2(data, lobs, steps = lobs, trend_only = FALSE,
  sensitivity = "low")
}
\arguments{
\item{data}{timseries like data to be modeled}

\item{lobs}{number of observations per day}

\item{steps}{number of data points to be predicted}

\item{trend_only}{logical TRUE/FALSE indicates if seasonal part should be included}

\item{sensitivity}{shows confidence interval has three values "low", "medium" and "high" (default is "low")}
}
\value{
prediction based on linear model and confidence intervals along with model reliability
}
\description{
Constructs linear model containing season and/or trend
}
\author{
Harutyun Khachatryan
}
