% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stationarity.R
\name{MultipleChange2}
\alias{MultipleChange2}
\title{Multiple stationarity change}
\usage{
MultipleChange2(intseries, Q = 5, robust = FALSE)
}
\arguments{
\item{intseries}{input data vector}

\item{Q}{maximal number of changepoints to be found}

\item{robust}{logical TRUE/FALSE indicating if robust method should be used}
}
\value{
most possible change point location index
}
\description{
Estimates possible stationarity point location in data which number at least is equal to Q
}
\examples{
x<- rnorm(1000)
MultipleChange2(x, Q = 5)
}
\author{
Harutyun Khachatryan
}
