% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STests.R
\name{MVTest2}
\alias{MVTest2}
\title{Performs mean and variance change test}
\usage{
MVTest2(itseries, ind, lobs)
}
\arguments{
\item{itseries}{time series like vector}

\item{ind}{index around which test is performed}

\item{lobs}{length of window around ind index}
}
\value{
0 when nor mean neither variance change is detected,
0.05 when mean or variance change is detected,
0.1 when both mean and variance change is detected
}
\description{
Does mean and variance change test around index ind of time series with window length lobs
}
\author{
Harutyun Khachatryan, Tigran Khachikyan
}
