% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodicity.R
\name{NRel2}
\alias{NRel2}
\title{Checks periodicity and gives reliability for having periodicity}
\usage{
NRel2(itseries, obsi)
}
\arguments{
\item{itseries}{timeseries like vector of values}

\item{obsi}{number of steps of possible periodicity to be checked}
}
\value{
vector of 2 elements containing:
a) input value obsi
b) reliability value spans from zero to unity
}
\description{
Performs different tests including ACF to detect periodicity in data
}
\author{
Harutyun Khachatryan, Tigran Khachikyan
}
