% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divider.R
\name{OneDayProj2}
\alias{OneDayProj2}
\title{Projects timestamps to one day, i.e. last appeared day in given vector of timestamps}
\usage{
OneDayProj2(time_stamps, ch_ind = NA)
}
\arguments{
\item{time_stamps}{original timestamps}

\item{ch_ind}{integer value of index at which concept changes last occur}
}
\value{
one day projected timestamps, i.e. some values of timestamps is changed
}
\description{
Takes original timestamps and projects them all to last day appeared in timestamps
}
\author{
Harutyun Khachatryan
}
