% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicter.R
\name{Predicter2}
\alias{Predicter2}
\title{Estimates prediction based dynamic thresholds}
\usage{
Predicter2(norm_data, norm_times, is_positive = TRUE, trend_only = FALSE,
  psense = "low")
}
\arguments{
\item{norm_data}{timeseries like observed with fixed time step values}

\item{norm_times}{timeseries observation times given as vector of POSIXct}

\item{is_positive}{logical TRUE/FALSE indicating that data non-negative or not}

\item{trend_only}{logical TRUE/FALSE indictaing if only trend will be predicted}

\item{psense}{shows confidence interval has three values "low", "medium" and "high" (default is "low")}
}
\value{
prediction dates along with lower, upper limits for data and reliability
}
\description{
Develops dynamics thresholds using simple lienear model with trend and season
}
\author{
Harutyun Khachatryan
}
