% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stationarity.R
\name{SingleChange2}
\alias{SingleChange2}
\title{Single stationarity change}
\usage{
SingleChange2(data)
}
\arguments{
\item{data}{input data vector}
}
\value{
most possible change point location index
}
\description{
Estimates single most possible stationarity point location in data
}
\examples{
x<- rnorm(1000)
SingleChange2(x)
}
\author{
Harutyun Khachatryan
}
