% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stationarity.R
\name{StatLen2}
\alias{StatLen2}
\title{Stationarity checker and divider}
\usage{
StatLen2(itseries, iobs, xhint = FALSE, xper = FALSE, nchanges = NA)
}
\arguments{
\item{itseries}{time series like vector}

\item{iobs}{length of window around ind index}

\item{xhint}{logical TRUE/FALSE indicating if data is highly unstable (default value is FALSE)}

\item{xper}{logical TRUE/FALSE indicating if data is periodic (default value is FALSE)}

\item{nchanges}{number of possible changes in data (default value is NA,
in that case nchanges is equal to one third of number of days in data)}
}
\description{
Finds all stationar periods of data. Maximal number of intervals is nchanges
}
\author{
Harutyun Khachatryan
}
