% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicter.R
\name{StepY2}
\alias{StepY2}
\title{Predicts mean value}
\usage{
StepY2(x, sensitivity = "low")
}
\arguments{
\item{x}{values to be predict}

\item{sensitivity}{shows confidence interval has three values "low", "medium" and "high" (default is "low")}
}
\description{
Performs mean value prediction and return confidence interval for it
}
\author{
Harutyun Khachatryan
}
