% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STests.R
\name{STest2}
\alias{STest2}
\title{Checks if any trend exists in data itseries}
\usage{
STest2(itseries)
}
\arguments{
\item{itseries}{time series like vector}
}
\value{
0 when nor mean neither variance change is detected,
0.1 when both mean and variance change is detected
}
\description{
Peroforms Spearman correlation test betwwen indeces and values of vector, if correlation exists then it must be some trend
}
\author{
Harutyun Khachatryan, Tigran Khachikyan
}
