% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicter.R
\name{YeoJohn2}
\alias{YeoJohn2}
\title{Performs Yeo-Johnson power transformation on a vector}
\usage{
YeoJohn2(x, lambda = 0)
}
\arguments{
\item{x}{vector or timeseries to be transformed}

\item{lambda}{transformation parameter obtained from BoxCox.lambda method of forecast package}
}
\value{
transformed values of input vector
}
\description{
Does Yeo-Johnson transformation on a vector with parameter lambda
}
\author{
Harutyun Khachatryan
}
