% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicter.R
\name{YeoJohnInv2}
\alias{YeoJohnInv2}
\title{Performs Yeo-Johnson inverse power transformation on a vector}
\usage{
YeoJohnInv2(x, lambda = 0)
}
\arguments{
\item{x}{vector or timeseries to be transformed}

\item{lambda}{transformation parameter obtained from BoxCox.lambda method of forecast package}
}
\value{
inverse transformed values of input vector
}
\description{
Does invserse (i.e. back) Yeo-Johnson transformation on a vector with parameter lambda
}
\author{
Harutyun Khachatryan
}
